devices/ArduSubPilot/pyproject.toml
[tool.poetry]
name = "ArduSubPilot"
version = "0.1.0"
description = ""
authors = ["Michael Hoyer <michael.hoyer3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pymavlink = "^2.4.9"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
